move newEntityTrackerEntry
