--- src/main/java/com/anatawa12/fixRtm/asm/Types.kt
@@ -1,7 +1,0 @@
-package com.anatawa12.fixRtm.asm
-
-object Types {
-    val hooksInternalName = "com/anatawa12/fixRtm/HooksKt"
-    val hooksName = hooksInternalName.replace('/', '.')
-    val entityTrackerEntryInternalName = "net/minecraft/entity/EntityTrackerEntry"
-}
